+++ src/main/java/hu/u_szeged/inf/fog/simulator/iot/actuator/DisconnectFromNode.java
@@ -4,7 +4,7 @@ import hu.u_szeged.inf.fog.simulator.application.Application;
 import hu.u_szeged.inf.fog.simulator.iot.Station;
 
 public class DisconnectFromNode implements ActuatorEvent {
-    public static int counter = 0;
+    public static long counter = 0;
     public DisconnectFromNode(){}
 
     @Override
