+++ src/main/java/hu/u_szeged/inf/fog/simulator/iot/actuator/TimeoutEvent.java
@@ -4,7 +4,7 @@ import hu.mta.sztaki.lpds.cloud.simulator.Timed;
 import hu.u_szeged.inf.fog.simulator.iot.Station;
 
 public class TimeoutEvent implements ActuatorEvent {
-    public static int counter = 0;
+    public static long counter = 0;
     public TimeoutEvent(){}
 
     @Override
